FEAT: getAdminRepository 위치 수정
