+++ apps/api/src/module/shared/transaction/repository.provider.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { TransactionService } from '@src/module/shared/transaction/transaction.service';
-import { getAdminRepository } from '@src/module/backoffice/domain/login-entity';
+import { getAdminRepository } from '@src/module/backoffice/domain/admin.entity';
 
 @Injectable()
 export class RepositoryProvider {
